+++ NineZeroProject/Service/HTQuestionService.m
@@ -132,7 +132,7 @@
     return _questionList;
 }
 
-- (void)getQuestionDetailWithQuestionID:(NSUInteger)questionID callback:(HTQuestionCallback)callback {
+- (void)getQuestionDetailWithQuestionID:(uint64_t)questionID callback:(HTQuestionCallback)callback {
     NSDictionary *dict = @{@"question_id" : [NSString stringWithFormat:@"%ld", (unsigned long)questionID],
                            @"area_id" : AppDelegateInstance.cityCode,
                            @"user_id" : [[HTStorageManager sharedInstance] getUserID]};
@@ -146,7 +146,7 @@
     }];
 }
 
-- (void)verifyQuestion:(NSUInteger)questionID withAnswer:(NSString *)answer callback:(HTResponseCallback)callback {
+- (void)verifyQuestion:(uint64_t)questionID withAnswer:(NSString *)answer callback:(HTResponseCallback)callback {
     NSString *user_id = [[HTStorageManager sharedInstance] getUserID];
     if (user_id.length == 0) {
         callback(false, nil);
@@ -209,7 +209,7 @@
     }];
 }
 
-- (void)verifyQuestion:(NSUInteger)questionID withLocation:(CGPoint)location callback:(HTResponseCallback)callback {
+- (void)verifyQuestion:(uint64_t)questionID withLocation:(CGPoint)location callback:(HTResponseCallback)callback {
     NSDictionary *dataDict = @{
                                @"lng" : @(location.x),
                                @"lat" : @(location.y),
@@ -296,7 +296,7 @@
     }];
 }
 
-- (HTQuestion *)findQuestionInQuestionList:(NSUInteger)questionID {
+- (HTQuestion *)findQuestionInQuestionList:(uint64_t)questionID {
     __block HTQuestion *question;
     [_questionList enumerateObjectsUsingBlock:^(HTQuestion * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
         if (questionID == obj.questionID) {
